chore: Switch to a hard-coded list of test clients and pre-bake into the test image (#206)
- Use a hard-coded set of test clients for internal development
- Pre-bake the clients into the test image for faster integration tests

This includes a change to the way clients are created. For the production mode no clients are created. Clients need to be created via the Keycloak admin UI / API and then the secrets can be supplied to the client.

For dev and testcontainer mode 3 standard clients are created:

 - OFF: A client with full permissions (can manage users) that would typically be used by Product Opener
 - test_client: A client with a secret that can be used to support a standard OAuth code flow with backend exchange for a token
 - test_public_client: A client with no secret that could be used for a PKCE style login flow

All of the clients have wildcard http://* redirect URLs so not suitable for production.